requirements.txt
APScheduler==3.9.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
dnspython==2.2.1
idna==3.3
lxml==4.9.1
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
